requirements.txt
pyspark
apache-sedona
geopandas
osmnx
langchain
streamlit
duckdb
requests
pystac-client
rasterio
rasterstats
matplotlib
notebook
ipykernel
pyarrow